Refactor folder structure and fix compiler issues on macOS
