+++ Sources/Shared/Core/BlueprintLayoutAnimationType.swift
