+++ Sources/Shared/Core/VerticalBlueprintLayout.swift
