+++ Sources/macOS/Extensions/BlueprintLayout+macOS.swift
@@ -2,7 +2,7 @@ import Cocoa
 
 extension BlueprintLayout {
   open override func layoutAttributesForSupplementaryView(ofKind elementKind: NSCollectionView.SupplementaryElementKind, at indexPath: IndexPath) -> LayoutAttributes? {
-    let sectionAttributes = layoutAttributes[indexPath.section]
+    let sectionAttributes = cachedAttributes[indexPath.section]
     var layoutAttributesResult: LayoutAttributes? = nil
 
     switch elementKind {
